refactor deriving
