+++ morpheus-graphql/src/Data/Morpheus/Server/Deriving/Schema/TypeContent.hs
@@ -4,9 +4,11 @@
 
 module Data.Morpheus.Server.Deriving.Schema.TypeContent
   ( buildTypeContent,
+    insertTypeContent,
   )
 where
 
+import Data.Morpheus.Server.Deriving.Schema.Directive (deriveTypeDirectives)
 import Data.Morpheus.Server.Deriving.Schema.Enum
   ( buildEnumTypeContent,
   )
@@ -25,8 +27,15 @@ import Data.Morpheus.Server.Deriving.Utils
 import Data.Morpheus.Server.Deriving.Utils.Kinded
   ( CategoryValue (..),
   )
-import Data.Morpheus.Server.Types.GQLType (GQLType)
-import Data.Morpheus.Server.Types.SchemaT (SchemaT)
+import Data.Morpheus.Server.Types.GQLType
+  ( GQLType (..),
+    deriveFingerprint,
+    deriveTypename,
+  )
+import Data.Morpheus.Server.Types.SchemaT
+  ( SchemaT,
+    updateSchema,
+  )
 import Data.Morpheus.Types.Internal.AST
 
 buildTypeContent ::
@@ -37,3 +46,24 @@ buildTypeContent ::
 buildTypeContent scope cons | all isEmptyConstraint cons = buildEnumTypeContent scope (consName <$> cons)
 buildTypeContent scope [ConsRep {consFields}] = buildObjectTypeContent scope consFields
 buildTypeContent scope cons = buildUnionTypeContent scope cons
+
+insertTypeContent ::
+  (GQLType a, CategoryValue kind) =>
+  (f kind a -> SchemaT c (TypeContent TRUE kind CONST)) ->
+  f kind a ->
+  SchemaT c ()
+insertTypeContent f proxy =
+  updateSchema
+    (deriveFingerprint proxy)
+    deriveD
+    proxy
+  where
+    deriveD x = do
+      content <- f x
+      dirs <- deriveTypeDirectives proxy
+      pure $
+        TypeDefinition
+          (description proxy)
+          (deriveTypename proxy)
+          dirs
+          content
